Add some missing Q_OBJECT macros
