+++ src/3d/qgs3dsceneexporter.h
@@ -54,6 +54,8 @@ class QgsMeshEntity;
  */
 class _3D_EXPORT Qgs3DSceneExporter : public Qt3DCore::QEntity
 {
+    Q_OBJECT
+
   public:
     //! Constructor
     Qgs3DSceneExporter() { }
