+++ src/3d/qgspointcloudlayerchunkloader_p.h
@@ -53,6 +53,8 @@
  */
 class QgsPointCloudLayerChunkLoaderFactory : public QgsChunkLoaderFactory
 {
+    Q_OBJECT
+
   public:
 
     /**
@@ -87,6 +89,8 @@ class QgsPointCloudLayerChunkLoaderFactory : public QgsChunkLoaderFactory
  */
 class QgsPointCloudLayerChunkLoader : public QgsChunkLoader
 {
+    Q_OBJECT
+
   public:
 
     /**
