+++ src/3d/qgsshadowrenderingframegraph.h
@@ -55,6 +55,8 @@ class QgsPreviewQuad;
  */
 class QgsShadowRenderingFrameGraph : public Qt3DCore::QEntity
 {
+    Q_OBJECT
+
   public:
     //! Constructor
     QgsShadowRenderingFrameGraph( QSurface *surface, QSize s, Qt3DRender::QCamera *mainCamera, Qt3DCore::QEntity *root );
