+++ src/3d/qgsskyboxentity.h
@@ -77,6 +77,8 @@ class _3D_EXPORT QgsSkyboxEntity : public Qt3DCore::QEntity
  */
 class _3D_EXPORT QgsPanoramicSkyboxEntity : public QgsSkyboxEntity
 {
+    Q_OBJECT
+
   public:
     //! Construct a skybox from a panoramic 360 image
     QgsPanoramicSkyboxEntity( const QString &texturePath, Qt3DCore::QNode *parent = nullptr );
@@ -102,6 +104,8 @@ class _3D_EXPORT QgsPanoramicSkyboxEntity : public QgsSkyboxEntity
  */
 class _3D_EXPORT QgsCubeFacesSkyboxEntity : public QgsSkyboxEntity
 {
+    Q_OBJECT
+
   public:
     //! Constructs a skybox from 6 different images
     QgsCubeFacesSkyboxEntity( const QString &posX, const QString &posY, const QString &posZ, const QString &negX, const QString &negY, const QString &negZ, Qt3DCore::QNode *parent = nullptr );
