+++ src/3d/symbols/qgspointcloud3dsymbol_p.h
@@ -125,6 +125,8 @@ class QgsClassificationPointCloud3DSymbolHandler : public QgsPointCloud3DSymbolH
 
 class QgsPointCloud3DGeometry: public Qt3DRender::QGeometry
 {
+    Q_OBJECT
+
   public:
     QgsPointCloud3DGeometry( Qt3DCore::QNode *parent, unsigned int byteStride );
 
@@ -142,6 +144,8 @@ class QgsPointCloud3DGeometry: public Qt3DRender::QGeometry
 
 class QgsSingleColorPointCloud3DGeometry : public QgsPointCloud3DGeometry
 {
+    Q_OBJECT
+
   public:
     QgsSingleColorPointCloud3DGeometry( Qt3DCore::QNode *parent, const QgsPointCloud3DSymbolHandler::PointData &data, unsigned int byteStride );
 
@@ -151,6 +155,8 @@ class QgsSingleColorPointCloud3DGeometry : public QgsPointCloud3DGeometry
 
 class QgsColorRampPointCloud3DGeometry : public QgsPointCloud3DGeometry
 {
+    Q_OBJECT
+
   public:
     QgsColorRampPointCloud3DGeometry( Qt3DCore::QNode *parent, const QgsPointCloud3DSymbolHandler::PointData &data, unsigned int byteStride );
 
@@ -160,6 +166,8 @@ class QgsColorRampPointCloud3DGeometry : public QgsPointCloud3DGeometry
 
 class QgsRGBPointCloud3DGeometry : public QgsPointCloud3DGeometry
 {
+    Q_OBJECT
+
   public:
     QgsRGBPointCloud3DGeometry( Qt3DCore::QNode *parent, const QgsPointCloud3DSymbolHandler::PointData &data, unsigned int byteStride );
   private:
