+++ src/3d/terrain/quantizedmeshterraingenerator.h
@@ -25,6 +25,8 @@
  */
 class QuantizedMeshTerrainGenerator : public QgsTerrainGenerator
 {
+    Q_OBJECT
+
   public:
     QuantizedMeshTerrainGenerator();
